Adicionado as features 08, 09, 10, 11, 15

+++ src/Domain/DTOs/ProducContext/StockMovementDTO.cs
@@ -4,6 +4,6 @@ namespace dotnet_api_erp.src.Domain.DTOs.ProducContext
 {
     public class StockMovementDTO
     {
-        public record CreateStockMovementDTO(Guid ProductId,  Guid UserId, string Description, MovementType MovementType, int Quantity, decimal Quote);
+        public record CreateStockMovementDTO(Guid ProductId,  Guid UserId, string Description, MovementType MovementType, int Quantity, double Quote);
     }
 }
